+++ app/src/main/java/com/nick_sib/popularlibraries/navigation/BackButtonListener.kt
@@ -1,4 +1,4 @@
-package com.nick_sib.popularlibraries.ui
+package com.nick_sib.popularlibraries.navigation
 
 interface BackButtonListener {
     fun backPressed(): Boolean
