+++ app/src/main/java/com/nick_sib/popularlibraries/ui/theuser/TheUserPresenter.kt
@@ -1,16 +1,10 @@
 package com.nick_sib.popularlibraries.ui.theuser
 
-import android.util.Log
-import com.nick_sib.popularlibraries.repo.GithubUser
-import com.nick_sib.popularlibraries.repo.GithubUsersRepo
-import com.nick_sib.popularlibraries.ui.Screens
-import com.nick_sib.popularlibraries.ui.users.UsersPresenter
+import com.nick_sib.popularlibraries.model.GithubUsersRepo
 import moxy.MvpPresenter
-import ru.terrakok.cicerone.Router
 
 class TheUserPresenter (
     private val usersRepo: GithubUsersRepo,
-    private val router: Router,
 ) : MvpPresenter<TheUserView>() {
 
     override fun onFirstViewAttach() {
@@ -22,8 +16,4 @@ class TheUserPresenter (
     fun theUserData(userIndex: Int) = usersRepo.getUser(userIndex)
 
 
-    fun backPressed(): Boolean {
-        router.exit()
-        return true
-    }
 }
